Conflicts:
	programs/oeis/000/A000566.asm
	programs/oeis/051/A051624.asm
	programs/oeis/051/A051872.asm
	programs/oeis/100/A100040.asm
	programs/oeis/100/A100041.asm
	programs/oeis/132/A132380.asm
	programs/oeis/135/A135706.asm
	programs/oeis/152/A152773.asm
	programs/oeis/152/A152965.asm
	programs/oeis/153/A153784.asm
	programs/oeis/153/A153785.asm
	programs/oeis/167/A167030.asm
	programs/oeis/190/A190816.asm
	programs/oeis/192/A192136.asm
	programs/oeis/300/A300404.asm

+++ src/memory.hpp
@@ -20,13 +20,13 @@ public:
 
   void clear();
 
-  void clear( number_t start, size_t length );
+  void clear( number_t start, int64_t length );
 
-  Memory fragment( number_t start, size_t length, bool normalize ) const;
+  Memory fragment( number_t start, int64_t length, bool normalize ) const;
 
   size_t approximate_size() const;
 
-  bool is_less( const Memory &m, size_t length ) const;
+  bool is_less( const Memory &m, int64_t length ) const;
 
   bool operator==( const Memory &m ) const;
